Rename arrow-heads to arrow.models
